Minor cleanup.

+++ main.m
@@ -178,7 +178,6 @@ int main(int argc, char *argv[])
         exit(2);
     }
 
-    // TODO (2005-07-27): Maybe add a flag to test whether the file has Objective-C segments, and return a different exit code.
     if (optind < argc) {
         NSString *arg, *executablePath;
 
@@ -225,7 +224,7 @@ int main(int argc, char *argv[])
             {
                 CDFatFile *fat = file;
                 NSArray *a1;
-                unsigned int count, index;
+                NSUInteger count, index;
 
                 a1 = [fat arches];
                 count = [a1 count];
@@ -245,7 +244,6 @@ int main(int argc, char *argv[])
             [classDump setTargetArch:targetArch];
             [classDump setExecutablePath:[executablePath stringByDeletingLastPathComponent]];
 
-            // NO, we have the file already: file
             if ([classDump loadFile:file]) {
 #if 0
                 [classDump showHeader];
